+++ src/AbyssIrc.Server/Program.cs
@@ -75,6 +75,12 @@ class Program
         }
 
 
+        if (Environment.GetEnvironmentVariable("ABYSS_ENABLE_DEBUG") != null)
+        {
+            options.EnableDebug = true;
+        }
+
+
         _hostBuilder = Host.CreateApplicationBuilder(args);
 
         if (string.IsNullOrWhiteSpace(options?.RootDirectory))
@@ -117,11 +123,29 @@ class Program
         if (options.EnableDebug)
         {
             loggingConfig.MinimumLevel.Debug();
+
+            // Additional log file for specific logs (e.g., TCP/Network)
+            loggingConfig.WriteTo.Logger(
+                lc => lc
+                    .MinimumLevel.Debug()
+                    .WriteTo.File(
+                        formatter: new CompactJsonFormatter(),
+                        path: Path.Combine(_directoriesConfig[DirectoryType.Logs], "network_debug_.log"),
+                        rollingInterval: RollingInterval.Day
+                    )
+                    // Filter to include only logs from specific namespaces or with specific properties
+                    .Filter.ByIncludingOnly(
+                        e =>
+                            e.Properties.ContainsKey("SourceContext") &&
+                            e.Properties["SourceContext"].ToString().Contains("Tcp") || e.Properties.ContainsKey("SourceContext").ToString().Contains("TcpService")
+                    )
+            );
         }
         else
         {
             loggingConfig.MinimumLevel.Information();
         }
+
 
         Log.Logger = loggingConfig.CreateLogger();
 
@@ -129,6 +153,7 @@ class Program
             .RegisterIrcCommandListener<QuitMessageHandler>(new QuitCommand())
             .RegisterIrcCommandListener<NickUserHandler>(new UserCommand())
             .RegisterIrcCommandListener<NickUserHandler>(new NickCommand())
+            .RegisterIrcCommandListener<NickUserHandler>(new IsonCommand())
             .RegisterIrcCommandListener<PingPongHandler>(new PingCommand())
             .RegisterIrcCommandListener<PingPongHandler>(new PongCommand())
             .RegisterIrcCommandListener<PrivMsgHandler>(new PrivMsgCommand());
@@ -146,7 +171,9 @@ class Program
             .RegisterIrcCommand(new PongCommand())
             .RegisterIrcCommand(new PrivMsgCommand())
             .RegisterIrcCommand(new ModeCommand())
-            .RegisterIrcCommand(new QuitCommand());
+            .RegisterIrcCommand(new QuitCommand())
+            .RegisterIrcCommand(new IsonCommand())
+            ;
 
 
         // Register handlers
@@ -180,6 +207,7 @@ class Program
             .RegisterScriptModule<EventsModule>()
             .RegisterScriptModule<SchedulerModule>()
             .RegisterScriptModule<IrcManagerModule>()
+            .RegisterScriptModule<VariableModule>()
             ;
 
 
@@ -220,6 +248,6 @@ class Program
         var version = assembly.GetName().Version;
 
         Console.WriteLine(reader.ReadToEnd());
-        Console.WriteLine($"Version: {version}");
+        Console.WriteLine($"  >> Version: {version}");
     }
 }