+++ src/AbyssIrc.Server/Servers/IrcTcpServer.cs
@@ -4,6 +4,7 @@ using AbyssIrc.Server.Data.Events.Client;
 using AbyssIrc.Server.Interfaces.Services;
 using AbyssIrc.Server.Interfaces.Services.System;
 using AbyssIrc.Server.Servers.Session;
+using AbyssIrc.Server.Types;
 using AbyssIrc.Signals.Interfaces.Services;
 using NetCoreServer;
 
@@ -25,7 +26,13 @@ public class IrcTcpServer : TcpServer
         _ircTcpServer = tcpService;
         _signalService = signalService;
         _sessionManagerService = sessionManagerService;
+
+        OptionNoDelay = true;
+        OptionReceiveBufferSize = 8192;
+        OptionSendBufferSize = 8192;
     }
+
+
 
     protected override TcpSession CreateSession()
     {
@@ -34,7 +41,7 @@ public class IrcTcpServer : TcpServer
 
     public async Task DispatchMessageAsync(string id, string message)
     {
-        _ircTcpServer.ParseCommandAsync(id, message);
+        _ircTcpServer.ParseCommandAsync(TcpServerType.Plain, id, message);
     }
 
     public async void ClientConnected(string id, string endPoint)